Move auth pages in auth folder
